+++ src/main/java/com/example/hw_5/service/FeedbackService.java
@@ -5,12 +5,12 @@ import com.example.hw_5.entity.Feedback;
 import java.util.List;
 
 public interface FeedbackService {
-    List<Feedback> getAllByRestaurantID(Integer id);
-    Double getAverageRatingByRestaurantID(Integer restaurantID);
-    String getFeedbackTextByID(Integer id);
+    List<Feedback> getAllByRestaurantID(Long id);
+    Double getAverageRatingByRestaurantID(Long restaurantID);
+    String getFeedbackTextByID(Long id);
 
     void addNewFeedback(Feedback feedback);
-    void changeFeedbackByID(Integer feedbackID, String newFeedback, Integer newRating);
+    void changeFeedbackByID(Long feedbackID, String newFeedback, Integer newRating);
 
-    void deleteFeedbackByRestaurantId(Integer id);
+    void deleteFeedbackByRestaurantId(Long id);
 }
